Refactor driver exception conversion tests
