+++ tests/Driver/AbstractOracleDriverTest.php
@@ -5,6 +5,8 @@ namespace Doctrine\DBAL\Tests\Driver;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Driver;
 use Doctrine\DBAL\Driver\AbstractOracleDriver;
+use Doctrine\DBAL\Driver\API\ExceptionConverter;
+use Doctrine\DBAL\Driver\API\OCI;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Platforms\OraclePlatform;
 use Doctrine\DBAL\Schema\AbstractSchemaManager;
@@ -27,43 +29,8 @@ class AbstractOracleDriverTest extends AbstractDriverTest
         return new OracleSchemaManager($connection);
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    protected static function getExceptionConversionData(): array
+    protected function createExceptionConverter(): ExceptionConverter
     {
-        return [
-            self::EXCEPTION_CONNECTION => [
-                [1017],
-                [12545],
-            ],
-            self::EXCEPTION_FOREIGN_KEY_CONSTRAINT_VIOLATION => [
-                [2292],
-            ],
-            self::EXCEPTION_INVALID_FIELD_NAME => [
-                [904],
-            ],
-            self::EXCEPTION_NON_UNIQUE_FIELD_NAME => [
-                [918],
-                [960],
-            ],
-            self::EXCEPTION_NOT_NULL_CONSTRAINT_VIOLATION => [
-                [1400],
-            ],
-            self::EXCEPTION_SYNTAX_ERROR => [
-                [923],
-            ],
-            self::EXCEPTION_TABLE_EXISTS => [
-                [955],
-            ],
-            self::EXCEPTION_TABLE_NOT_FOUND => [
-                [942],
-            ],
-            self::EXCEPTION_UNIQUE_CONSTRAINT_VIOLATION => [
-                [1],
-                [2299],
-                [38911],
-            ],
-        ];
+        return new OCI\ExceptionConverter();
     }
 }
